[None][test] unify VideoMME EPD tests into TestVideoMMEEPD harness
Move the encode/prefill-decode (E/PD) VideoMME multimodal tests out of
test_disaggregated_serving.py into a dedicated test_epd_disagg_multimodal.py with
a unified class-based harness: TestVideoMMEEPD(LlmapiAccuracyTestHarness) plus a
frozen EPDVariant dataclass whose classmethods (qwen3vl_2b, nano_omni_fp8,
nano_omni_nvfp4) build immutable per-variant configs (no post-construction
mutation). Lift the shared sampling params and no-thinking evaluator kwargs to
class constants, keep one parametrized test, and drop the config meta-test.
Update the three test-list references to the new file/class path.
